requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.3.0
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
jsonpath==0.82
openpyxl==3.0.9
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
PyYAML==6.0
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
